removed old testing file I was unaware of; stemmer at 99.66%

+++ stemmers/porter2_stemmer.h
@@ -63,8 +63,7 @@ namespace Porter2Stemmer
 
         bool isValidLIEnding(char ch);
 
-        bool containsVowel(const std::string & word, size_t start,
-                size_t end);
+        bool containsVowel(const std::string & word, int start, int end);
     }
 }
